Update node reward pop-up
