+++ Assets/Scripts/Base/UI/HUD/UI_LevelResult.cs
@@ -9,7 +9,7 @@ public enum LevelResultType
     OUT_OF_TIME
 }
 
-// Temporary UI for level result, to integrated with UI_Manager System
+// Temporary UI for level result, to be integrated with UI_Manager System
 public class UI_LevelResult : MonoBehaviour
 {
     [SerializeField] GameObject m_LevelResultPanel;
@@ -18,7 +18,6 @@ public class UI_LevelResult : MonoBehaviour
 
     private void Awake()
     {
-        m_ReturnButton.onClick.AddListener(ReturnFromLevel);
         m_LevelResultPanel.SetActive(false);
         GlobalEvents.Level.LevelEndEvent += OnLevelEnd;
     }
@@ -38,11 +37,13 @@ public class UI_LevelResult : MonoBehaviour
             _ => "???"
         };
         m_LevelResultPanel.SetActive(true);
+        m_ReturnButton.onClick.AddListener(ReturnFromLevel);
     }
     
     public void ReturnFromLevel()
     {
         m_LevelResultPanel.SetActive(false);
         GlobalEvents.Level.ReturnFromLevelEvent?.Invoke();
+        m_ReturnButton.onClick.RemoveListener(ReturnFromLevel);
     }
 }
