+++ Assets/Scripts/Level/Nodes/RewardNode.cs
@@ -4,6 +4,7 @@ public class RewardNode : NodeInternal
 {
     // Main currency (STC)
     [SerializeField] private int m_goldReward;
+    public int GoldReward => m_goldReward;
     
     // Secondary currency (STC)
     [SerializeField] private int m_manaStoneReward;
